refactor: [Artist] delete test fixtures
