+++ src/v2/Apps/Artist2/Routes/AuctionResults/__tests__/Artist2AuctionResults.jest.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable jest/no-done-callback */
 import { AuctionResults_Test_QueryRawResponse } from "v2/__generated__/AuctionResults_Test_Query.graphql"
-import { AuctionResultsFixture } from "v2/Apps/__tests__/Fixtures/Artist/Routes/AuctionResultsFixture"
 import { AuctionResultsRouteFragmentContainer as AuctionResultsRoute } from "v2/Apps/Artist2/Routes/AuctionResults/Artist2AuctionResultsRoute"
 import { MockBoot, renderRelayTree } from "v2/DevTools"
 import { ReactWrapper } from "enzyme"
@@ -421,3 +420,266 @@ describe("AuctionResults", () => {
     })
   })
 })
+
+const AuctionResultsFixture: AuctionResults_Test_QueryRawResponse = {
+  artist: {
+    id: "QXJ0aXN0OnBhYmxvLXBpY2Fzc28=",
+    slug: "pablo-picasso",
+    name: "Pablo Picasso",
+    auctionResultsConnection: {
+      pageInfo: { hasNextPage: true, endCursor: "YXJyYXljb25uZWN0aW9uOjk=" },
+      pageCursors: {
+        around: [
+          { cursor: "YXJyYXljb25uZWN0aW9uOi0x", page: 1, isCurrent: true },
+          { cursor: "YXJyYXljb25uZWN0aW9uOjk=", page: 2, isCurrent: false },
+          { cursor: "YXJyYXljb25uZWN0aW9uOjE5", page: 3, isCurrent: false },
+          { cursor: "YXJyYXljb25uZWN0aW9uOjI5", page: 4, isCurrent: false },
+        ],
+        first: null,
+        last: {
+          cursor: "YXJyYXljb25uZWN0aW9uOjgxOQ==",
+          page: 83,
+          isCurrent: false,
+        },
+        previous: null,
+      },
+      totalCount: 830,
+      createdYearRange: {
+        startAt: 1880,
+        endAt: 1973,
+      },
+      edges: [
+        {
+          node: {
+            title: "Oiseau fantastique",
+            dimension_text: "27.2 x 21.1 cm",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/xACxJ_uIHApai3JP9odtZg/thumbnail.jpg",
+              },
+            },
+            description:
+              "Pablo Picasso (1881-1973)\n\nOiseau fantastique\n\nfelt-tip pen on paper\n\n10 ¾ x 8 3/8 in. (27.2 x 21.1 cm.)\n\nDrawn in 1952",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$40,000 - 60,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDk2Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Mandoline sur une table",
+            dimension_text: "82.2 x 100.4 cm",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/lmY_wowdeGi__ZtKVHV8Dw/thumbnail.jpg",
+              },
+            },
+            description:
+              "Pablo Picasso (1881-1973)\n\nMandoline sur une table\n\nsigned 'Picasso' (lower right)\n\noil on canvas\n\n32 3/8 x 39 ½ in. (82.2 x 100.4 cm.)\n\nPainted in 1922",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00-05:00",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$4,000,000 - 6,000,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDc3Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Tête d'homme",
+            dimension_text: "51.2 x 34.2 cm",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/AI6P5qi0Xq7Efs9d6HMt4A/thumbnail.jpg",
+              },
+            },
+            description:
+              "Pablo Picasso (1881-1973)\n\nTête d'homme\n\ninscribed and dated 'Boisgeloup 10 juillet XXXIV' (upper right)\n\npencil on paper\n\n20 1/8 x 13 ½ in. (51.2 x 34.2 cm.)\n\nDrawn in Boisgeloup on 10 July 1934",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$500,000 - 700,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDc1Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Picador et taureau (A.R. 197)",
+            dimension_text: "23.5 cm.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/B3EtIMtH8XnDmt1KBD6VhQ/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\nPicador et taureau _(A.R. 197)_\n\ndated '25.9.53.' (in reverse; lower center); stamped and numbered 'Madoura Plein Feu/Empreinte Originale de Picasso/29/200' (underneath)\n\nwhite earthenware ceramic plate with black oxide, colored engobe and glaze\n\nDiameter: 9 ¼ in. (23.5 cm.)\n\nConceived on 25 September 1953 and executed in a numbered edition of 200",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$3,500 - 4,500" },
+            id: "QXVjdGlvblJlc3VsdDoyNDY4Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Colombe à la lucarne (A.R. 78)",
+            dimension_text: "15 3/8 in.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/rLyB6jNe0lQ8fF6EEQ61wg/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\nColombe à la lucarne _(A.R. 78)_\n\nstamped, marked and numbered 'Madoura Plein Feu/Edition Picasso/193/200/Edition Picasso/Madoura' (underneath)\n\nwhite earthenware ceramic plate, partially engraved, with colored engobe and glaze\n\nLength: 15 3/8 in. (39 cm.)\n\nConceived in 1949 and executed in a numbered edition of 200",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$6,000 - 8,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDY3Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Scène de plage (A.R. 391)",
+            dimension_text: "10 in.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/46t-8KytTjCwYPw17E7U6w/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\nScène de plage _(A.R. 391)_\n\nstamped 'Madoura Plein Feu/Empreinte Originale de Picasso' (on the reverse)\n\nwhite earthenware ceramic plaque with black engobe and white glaze\n\nDiameter: 10 in. (25.3 cm.)\n\nConceived in 1956 and executed in an edition of 450",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$2,500 - 3,500" },
+            id: "QXVjdGlvblJlc3VsdDoyNDY2Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Tête de femme couronnée de fleurs (A.R. 236)",
+            dimension_text: "9 1/8 in.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/eivcrcx7PVnvmKZzOQosXA/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\nTête de femme couronnée de fleurs _(A.R. 236)_\n\nsigned and dated 'Picasso/20.3.54.' (on the side); numbered '18/50/6.' (underneath)\n\nwhite earthenware ceramic vase, partially engraved, with white engobe\n\nHeight: 9 1/8 in. (23 cm.)\n\nConceived on 20 March 1954 and executed in a numbered edition of 50",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$18,000 - 25,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDY1Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Femme (A.R. 297)",
+            dimension_text: "13 in.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/-ZlQnxE8T8MsVRGjSSwaXw/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\nFemme _(A.R. 297)_\n\nstamped 'Edition Picasso/Madoura Plein Feu/Edition Picasso' (underneath)\n\nwhite earthenware ceramic pitcher with colored engobe and glaze\n\nHeight: 13 in. (33 cm.)\n\nConceived in 1955 and executed in an edition of 100",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$10,000 - 15,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDY0Nw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Taureau dans l'arène (A.R. 80)",
+            dimension_text: "37.6 cm.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/3nGESp60mCg0xygJ4bvjcA/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\n_Taureau dans l'arène _(A.R. 80)\n\nstamped and marked 'Madoura Plein Feu/D'Après Picasso/Edition Picasso /GR' (underneath)\n\nwhite earthenware ceramic plate, partially engraved, with colored engobe and glaze\n\nLength: 14 ¾ in. (37.6 cm.)\n\nConceived in 1948 and executed in an edition of 450",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: false,
+            estimate: { display: "$7,000 - 10,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDYzNw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+        {
+          node: {
+            title: "Visage de femme (A.R. 192)",
+            dimension_text: "13 3/8 in.",
+            organization: "Christie's",
+            images: {
+              thumbnail: {
+                url:
+                  "https://d32dm0rphc51dk.cloudfront.net/Db93v-hdsJjCV6XHFUfn2g/thumbnail.jpg",
+              },
+            },
+            description:
+              "PABLO PICASSO (1881-1973)\n\nVisage de femme _(A.R. 192)_\n\nstamped, marked and numbered 'Edition Picasso/Madoura Plein Feu/Edition Picasso/99/200/Madoura' (underneath)\n\nwhite earthenware ceramic pitcher, partially engraved, with colored engobe and glaze\n\nHeight: 13 3/8 in. (34 cm.)\n\nConceived on 7 July 1953 and executed in a numbered edition of 200",
+            date_text: "1881-1973",
+            saleDate: "2020-02-05T19:00:00Z",
+            price_realized: { display: "$NaN", cents_usd: 0 },
+            boughtIn: true,
+            estimate: { display: "$7,000 - 10,000" },
+            id: "QXVjdGlvblJlc3VsdDoyNDYyNw==",
+            mediumText: "oil on canvas",
+            categoryText: "Painting",
+          },
+        },
+      ],
+    },
+  },
+}
