--- src/v2/Apps/__tests__/Fixtures/Artist/Routes/ArticlesFixture.ts
@@ -1,218 +1,0 @@
-import { Articles_Test_QueryRawResponse } from "v2/__generated__/Articles_Test_Query.graphql"
-
-export const ArticlesFixture: Articles_Test_QueryRawResponse = {
-  artist: {
-    id: "QXJ0aXN0OnBhYmxvLXBpY2Fzc28=",
-    slug: "pablo-picasso",
-    articlesConnection: {
-      pageInfo: { hasNextPage: true, endCursor: "YXJyYXljb25uZWN0aW9uOjk=" },
-      pageCursors: {
-        around: [
-          { cursor: "YXJyYXljb25uZWN0aW9uOi0x", page: 1, isCurrent: true },
-          { cursor: "YXJyYXljb25uZWN0aW9uOjk=", page: 2, isCurrent: false },
-          { cursor: "YXJyYXljb25uZWN0aW9uOjE5", page: 3, isCurrent: false },
-          { cursor: "YXJyYXljb25uZWN0aW9uOjI5", page: 4, isCurrent: false },
-        ],
-        first: null,
-        last: null,
-        previous: null,
-      },
-      edges: [
-        {
-          node: {
-            href: "/article/artsy-editorial-us-china-trade-war-art-market",
-            thumbnail_title:
-              "What the U.S.–China Trade War Means for the Art Market",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjVhYmJmYzZlY2I0YzI3MTczNzcyOWE1YQ==",
-            },
-            published_at: "Jul 23rd, 2018",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2F0wcjPCKn9wQg_DRvywHDsQ%252FGettyImages-997234250-1200.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YjU2Mjk0OTMyMjliNDAwMmU0NmZhMjQ=",
-          },
-        },
-        {
-          node: {
-            href: "/article/artsy-editorial-picassos-muses-art-markets-darling",
-            thumbnail_title:
-              "Which of Picasso’s Muses Is the Art Market’s Darling?",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjVhYmJmYzZlY2I0YzI3MTczNzcyOWE1YQ==",
-            },
-            published_at: "Jun 22nd, 2018",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FXsyUiT-8FinACaBkvMXGhA%252Fpicassomag.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YjJiYmNiZmQ3YTNiMDAwNDI5MGM0Mzc=",
-          },
-        },
-        {
-          node: {
-            href: "/article/artsy-editorial-2018-picassos-billion-dollar-year",
-            thumbnail_title: "Why 2018 Could Be Picasso’s Billion Dollar Year",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU4Yjk5NjM4YTA5YTY3MzJjZWUwNGFiMw==",
-            },
-            published_at: "Apr 10th, 2018",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2F8VtjX1i3z4kT-Nbw_lBdbQ%252Fcustom-Custom_Size___Picasso%252C%2BLe%2BRepos%2B%25281932%2529%2B%25281%2529.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YWNiZDM3ZjVmMDFmMDAwMmU3YzM4Mzg=",
-          },
-        },
-        {
-          node: {
-            href:
-              "/article/artsy-editorial-1127-million-picasso-spending-spree-buoys-big-sales-christies-sothebys-london",
-            thumbnail_title:
-              "A £112.7 Million Picasso Spending Spree Buoys Big Sales at Christie’s and Sotheby’s in London",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Mar 1st, 2018",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FAXNlAM9edEcFLSh932x4GA%252FCMD_6770.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YTk3MzI1M2M1M2RkODAwMjk4ODY4MDY=",
-          },
-        },
-        {
-          node: {
-            href:
-              "/article/artsy-editorial-hidden-details-uncovered-beneath-surface-blue-period-picasso",
-            thumbnail_title:
-              "Hidden Details Uncovered beneath Surface of Blue Period Picasso",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Feb 20th, 2018",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FPCRy4AOwb6tU02iHJldQBQ%252Fxrf-set-up-with-people-la-misereuse-accroupie.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YThjNDcwMTkwZjVlZjAwMjlkZDc1YjA=",
-          },
-        },
-        {
-          node: {
-            href:
-              "/article/artsy-editorial-13-famous-artist-couples-massive-gender-pay-gaps",
-            thumbnail_title:
-              "13 Famous Artist Couples’ Massive Gender Pay Gaps",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Dec 29th, 2017",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FAHDPB6uTw16QzZ2FAyW-ag%252Fpicasso%2Bcouple.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YTNkNTY2MTA2NzMxMTAwMzMyYmI5OGQ=",
-          },
-        },
-        {
-          node: {
-            href:
-              "/article/artsy-editorial-asked-expert-interpret-famous-artists-handwriting",
-            thumbnail_title:
-              "We Asked an Expert to Interpret Famous Artists’ Handwriting",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Dec 21st, 2017",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FnHaY68AdX7Y-mmVoJzhk_w%252Fgeorgia%2Bhandwriting.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YTM5NzBhNzcyMDM1OTAwMWYzNjAwYjU=",
-          },
-        },
-        {
-          node: {
-            href:
-              "/article/artsy-editorial-emotional-turmoil-picassos-blue-period",
-            thumbnail_title:
-              "The Emotional Turmoil behind Picasso’s Blue Period",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Dec 13th, 2017",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FSCqsGjpeTT4wy-33OLKUkA%252Fpicasso%2Bblue%2Bthumb.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1YTJlYjUyNmUyOWM3MzAwMWYxOTkxMTA=",
-          },
-        },
-        {
-          node: {
-            href: "/article/artsy-editorial-picasso-invented-abstract-painting",
-            thumbnail_title: "When Picasso Almost Invented Abstract Painting",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Oct 27th, 2017",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FjLgdIxwTR61feM3qCOQ7XQ%252Fpicasso%2Bthumb.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1OWYwZDYzZDdkZTRiMDAwMWRjYmFkNDg=",
-          },
-        },
-        {
-          node: {
-            href:
-              "/article/artsy-editorial-jean-michel-basquiat-georgia-okeeffe-8-artists-style-icons",
-            thumbnail_title:
-              "From Jean-Michel Basquiat to Georgia O’Keeffe, 8 Artists Who Are Style Icons",
-            author: {
-              name: "Artsy Editorial",
-              id: "QXV0aG9yOjU5OGI1NDRkMmE4OTNhNTk1YzQxYzRkMQ==",
-            },
-            published_at: "Oct 18th, 2017",
-            thumbnail_image: {
-              resized: {
-                url:
-                  "https://d7hftxdivxxvm.cloudfront.net?resize_to=width&width=300&quality=80&src=https%3A%2F%2Fartsy-media-uploads.s3.amazonaws.com%2FjplLkTPL-P0BY4Zw6QF6WA%252Fstyle%2Bthumb.jpg",
-              },
-            },
-            id: "QXJ0aWNsZTo1OWU2M2Y5YTRhNzZlMzAwMWQyNTY1OTg=",
-          },
-        },
-      ],
-    },
-  },
-}
