--- src/v2/Apps/__tests__/Fixtures/Artist/Routes/CVFixture.tsx
@@ -1,583 +1,0 @@
-import { CV_Test_QueryRawResponse } from "v2/__generated__/CV_Test_Query.graphql"
-
-export const CVFixture: CV_Test_QueryRawResponse = {
-  viewer: {
-    artist_soloShows: {
-      slug: "pablo-picasso",
-      showsConnection: {
-        pageInfo: { hasNextPage: true, endCursor: "YXJyYXljb25uZWN0aW9uOjk=" },
-        edges: [
-          {
-            node: {
-              id: "U2hvdzp3YWRhLWdhcm91LXRva3lvLXBpY2Fzc28tcHJpbnRzLTE=",
-              partner: {
-                __typename: "Partner",
-                name: "Wada Garou Tokyo",
-                href: "/wada-garou-tokyo",
-                id: "UGFydG5lcjp3YWRhLWdhcm91LXRva3lv",
-              },
-              name: "Picasso Prints",
-              start_at: "2018",
-              city: null,
-              href: "/show/wada-garou-tokyo-picasso-prints-1",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjA=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpiYWlsbHktZ2FsbGVyeS1pdHMtb25seS1hYm91dC1jZXJhbWljcw==",
-              partner: {
-                __typename: "Partner",
-                name: "BAILLY GALLERY",
-                href: "/bailly-gallery",
-                id: "UGFydG5lcjpiYWlsbHktZ2FsbGVyeQ==",
-              },
-              name: "It's only about Ceramics",
-              start_at: "2018",
-              city: null,
-              href: "/show/bailly-gallery-its-only-about-ceramics",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjE=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpiYWlsbHktZ2FsbGVyeS1zb2xvLXNob3ctcGFibG8tcGljYXNzbw==",
-              partner: {
-                __typename: "Partner",
-                name: "BAILLY GALLERY",
-                href: "/bailly-gallery",
-                id: "UGFydG5lcjpiYWlsbHktZ2FsbGVyeQ==",
-              },
-              name: "Solo Show Pablo Picasso",
-              start_at: "2018",
-              city: "Paris",
-              href: "/show/bailly-gallery-solo-show-pablo-picasso",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjI=",
-          },
-          {
-            node: {
-              id: "U2hvdzpoaWRkZW4tcGljYXNzby10aGUtbWFzdGVy",
-              partner: {
-                __typename: "Partner",
-                name: "Hidden",
-                href: "/hidden",
-                id: "UGFydG5lcjpoaWRkZW4=",
-              },
-              name: "Picasso - The Master",
-              start_at: "2018",
-              city: null,
-              href: "/show/hidden-picasso-the-master",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjM=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzptYXN0ZXJ3b3Jrcy1maW5lLWFydC12aXNhZ2UtdGhlLW1hbnktZmFjZXMtb2YtcGljYXNzb3MtY2VyYW1pY3M=",
-              partner: {
-                __typename: "Partner",
-                name: "Masterworks Fine Art",
-                href: "/masterworks-fine-art-gallery",
-                id: "UGFydG5lcjptYXN0ZXJ3b3Jrcy1maW5lLWFydA==",
-              },
-              name: "Visage: The many faces of Picasso's ceramics",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/masterworks-fine-art-visage-the-many-faces-of-picassos-ceramics",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjQ=",
-          },
-          {
-            node: {
-              id: "U2hvdzpnYWxsZXJpYS10ZWdhLXBhYmxvLXBpY2Fzc28=",
-              partner: {
-                __typename: "Partner",
-                name: "Galleria Tega",
-                href: "/galleria-tega",
-                id: "UGFydG5lcjpnYWxsZXJpYS10ZWdh",
-              },
-              name: "Pablo Picasso",
-              start_at: "2018",
-              city: "Milan",
-              href: "/show/galleria-tega-pablo-picasso",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjU=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzp0YXRlLW1vZGVybi10aGUtZXktZXhoaWJpdGlvbi1waWNhc3NvLTE5MzItbG92ZS1mYW1lLXRyYWdlZHk=",
-              partner: {
-                __typename: "Partner",
-                name: "Tate Modern",
-                href: "/tate",
-                id: "UGFydG5lcjp0YXRlLW1vZGVybg==",
-              },
-              name: "The EY Exhibition: PICASSO 1932 – Love, Fame, Tragedy",
-              start_at: "2018",
-              city: "London",
-              href:
-                "/show/tate-modern-the-ey-exhibition-picasso-1932-love-fame-tragedy",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjY=",
-          },
-          {
-            node: {
-              id: "U2hvdzphbGJlcnQtbWVyb2xhLWdhbGxlcnktcGljYXNzby0xMi0zNC01",
-              partner: {
-                __typename: "Partner",
-                name: "Albert Merola Gallery",
-                href: "/albert-merola-gallery",
-                id: "UGFydG5lcjphbGJlcnQtbWVyb2xhLWdhbGxlcnk=",
-              },
-              name: "Picasso 1,2,3,4,5",
-              start_at: "2018",
-              city: "Provincetown",
-              href: null,
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjc=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzptdXNlZS1waWNhc3NvLXBhcmlzLXBpY2Fzc28tMTkzMi1hbm5lZS1lcm90aXF1ZQ==",
-              partner: {
-                __typename: "Partner",
-                name: "Musée Picasso Paris",
-                href: "/musee-picasso-paris",
-                id: "UGFydG5lcjptdXNlZS1waWNhc3NvLXBhcmlz",
-              },
-              name: "Picasso 1932. Année érotique",
-              start_at: "2017",
-              city: "Paris",
-              href: null,
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjg=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpjaHJpc3RvcGhlci1jbGFyay1maW5lLWFydC1waWNhc3NvLWluc3BpcmF0aW9uLW9mLXRoZS1tb2RlbC1hbmQtbXVzZQ==",
-              partner: {
-                __typename: "Partner",
-                name: "Christopher-Clark Fine Art",
-                href: "/christopher-clark-fine-art",
-                id: "UGFydG5lcjpjaHJpc3RvcGhlci1jbGFyay1maW5lLWFydA==",
-              },
-              name: "Picasso:  Inspiration of the Model and Muse",
-              start_at: "2017",
-              city: null,
-              href:
-                "/show/christopher-clark-fine-art-picasso-inspiration-of-the-model-and-muse",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjk=",
-          },
-        ],
-      },
-      id: "QXJ0aXN0OnBhYmxvLXBpY2Fzc28=",
-    },
-    artist_groupShows: {
-      slug: "pablo-picasso",
-      showsConnection: {
-        pageInfo: { hasNextPage: true, endCursor: "YXJyYXljb25uZWN0aW9uOjk=" },
-        edges: [
-          {
-            node: {
-              id:
-                "U2hvdzpsYXJzZW4tZ2FsbGVyeS0yMDE4LWxhcnNlbi1hcnQtYXVjdGlvbg==",
-              partner: {
-                __typename: "Partner",
-                name: "Larsen Gallery",
-                href: "/larsen-gallery",
-                id: "UGFydG5lcjpsYXJzZW4tZ2FsbGVyeQ==",
-              },
-              name: "2018 Larsen Art Auction",
-              start_at: "2018",
-              city: "Scottsdale",
-              href: "/show/larsen-gallery-2018-larsen-art-auction",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjA=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzptYXJ0aW4tbGF3cmVuY2UtZ2FsbGVyaWVzLW1vZGVybi1tYXN0ZXJzLTE=",
-              partner: {
-                __typename: "Partner",
-                name: "Martin Lawrence Galleries",
-                href: "/martin-lawrence-galleries",
-                id: "UGFydG5lcjptYXJ0aW4tbGF3cmVuY2UtZ2FsbGVyaWVz",
-              },
-              name: "Modern Masters",
-              start_at: "2018",
-              city: "New Orleans ",
-              href: "/show/martin-lawrence-galleries-modern-masters-1",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjE=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzptYWRkb3gtZ2FsbGVyeS1hdXR1bW4tY29udGVtcG9yYXJ5LWdzdGFhZC1zd2l0emVybGFuZA==",
-              partner: {
-                __typename: "Partner",
-                name: "Maddox Gallery",
-                href: "/maddox-gallery",
-                id: "UGFydG5lcjptYWRkb3gtZ2FsbGVyeQ==",
-              },
-              name: "Autumn Contemporary - Gstaad, Switzerland",
-              start_at: "2018",
-              city: "Gstaad",
-              href:
-                "/show/maddox-gallery-autumn-contemporary-gstaad-switzerland",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjI=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpsZXZpdG9uLWZpbmUtYXJ0LWFydHN5LWdyYW5kLW9wZW5pbmctYmx1ZS1jaGlwLWFydGlzdHM=",
-              partner: {
-                __typename: "Partner",
-                name: "Leviton Fine Art",
-                href: "/leviton-fine-art",
-                id: "UGFydG5lcjpsZXZpdG9uLWZpbmUtYXJ0",
-              },
-              name: "Artsy Grand Opening, Blue Chip Artists",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/leviton-fine-art-artsy-grand-opening-blue-chip-artists",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjM=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzp1bnRpdGxlZC1wcm9qZWN0cy0yMHRoLWNlbnR1cnktbWFzdGVycy1wcmludHMtYW5kLXBhaW50aW5n",
-              partner: {
-                __typename: "Partner",
-                name: "Untitled Projects",
-                href: "/untitled-projects",
-                id: "UGFydG5lcjp1bnRpdGxlZC1wcm9qZWN0cw==",
-              },
-              name: "20th Century Masters: Prints & Painting",
-              start_at: "2018",
-              city: "Los Angeles",
-              href:
-                "/show/untitled-projects-20th-century-masters-prints-and-painting",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjQ=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzp0aGUtbG9mdC1maW5lLWFydC0yNS1wZXJjZW50LW9mZi1zdG9yZS13aWRlLXNhbGU=",
-              partner: {
-                __typename: "Partner",
-                name: "The Loft Fine Art",
-                href: "/the-loft-fine-art",
-                id: "UGFydG5lcjp0aGUtbG9mdC1maW5lLWFydA==",
-              },
-              name: "25% Off Store-Wide Sale",
-              start_at: "2018",
-              city: null,
-              href: "/show/the-loft-fine-art-25-percent-off-store-wide-sale",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjU=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzphcnR3aXNlLW51bWJlci1hcnR3aXNldXAtdGltZXN0YW1wZWQtZXhoaWJpdGlvbi1wb3N0ZXJz",
-              partner: {
-                __typename: "Partner",
-                name: "ArtWise",
-                href: "/artwise",
-                id: "UGFydG5lcjphcnR3aXNl",
-              },
-              name: "#ArtWiseUP: Timestamped Exhibition Posters",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/artwise-number-artwiseup-timestamped-exhibition-posters",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjY=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpyb2dhbGxlcnktdHdvLWRheXMtb2Ytc2VwdGVtYmVyLWF1Y3Rpb25z",
-              partner: {
-                __typename: "Partner",
-                name: "RoGallery",
-                href: "/rogallery",
-                id: "UGFydG5lcjpyb2dhbGxlcnk=",
-              },
-              name: "Two Days of September Auctions",
-              start_at: "2018",
-              city: "Long Island City",
-              href: "/show/rogallery-two-days-of-september-auctions",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjc=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzptYXJsYm9yb3VnaC1nYWxsZXJ5LTEtcG9ydHJhaXRzLWluLXByaW50",
-              partner: {
-                __typename: "Partner",
-                name: "Marlborough Gallery",
-                href: "/marlborough-gallery",
-                id: "UGFydG5lcjptYXJsYm9yb3VnaC1nYWxsZXJ5LTE=",
-              },
-              name: "Portraits in Print",
-              start_at: "2018",
-              city: "Marlborough Graphics, New York",
-              href: "/show/marlborough-gallery-1-portraits-in-print",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjg=",
-          },
-          {
-            node: {
-              id: "U2hvdzphbGJlcnQtbWVyb2xhLWdhbGxlcnktaGVhZDJoZWFk",
-              partner: {
-                __typename: "Partner",
-                name: "Albert Merola Gallery",
-                href: "/albert-merola-gallery",
-                id: "UGFydG5lcjphbGJlcnQtbWVyb2xhLWdhbGxlcnk=",
-              },
-              name: "HEAD2HEAD",
-              start_at: "2018",
-              city: "Provincetown",
-              href: "/show/albert-merola-gallery-head2head",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjk=",
-          },
-        ],
-      },
-      id: "QXJ0aXN0OnBhYmxvLXBpY2Fzc28=",
-    },
-    artist_fairBooths: {
-      slug: "pablo-picasso",
-      showsConnection: {
-        pageInfo: { hasNextPage: true, endCursor: "YXJyYXljb25uZWN0aW9uOjk=" },
-        edges: [
-          {
-            node: {
-              id:
-                "U2hvdzpqb2huLXN6b2tlLTEtam9obi1zem9rZS1hdC1pZnBkYS1maW5lLWFydC1wcmludC1mYWlyLTIwMTg=",
-              partner: {
-                __typename: "Partner",
-                name: "John Szoke",
-                href: "/john-szoke-1",
-                id: "UGFydG5lcjpqb2huLXN6b2tlLTE=",
-              },
-              name: "John Szoke at IFPDA Fine Art Print Fair 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/john-szoke-1-john-szoke-at-ifpda-fine-art-print-fair-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjA=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpoYXJyaXMtc2NocmFuay1maW5lLXByaW50cy1oYXJyaXMtc2NocmFuay1maW5lLXByaW50cy1hdC1pZnBkYS1maW5lLWFydC1wcmludC1mYWlyLTIwMTg=",
-              partner: {
-                __typename: "Partner",
-                name: "Harris Schrank Fine Prints",
-                href: "/harris-schrank-fine-prints",
-                id: "UGFydG5lcjpoYXJyaXMtc2NocmFuay1maW5lLXByaW50cw==",
-              },
-              name:
-                "Harris Schrank Fine Prints at IFPDA Fine Art Print Fair 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/harris-schrank-fine-prints-harris-schrank-fine-prints-at-ifpda-fine-art-print-fair-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjE=",
-          },
-          {
-            node: {
-              id: "U2hvdzpkaWUtZ2FsZXJpZS1kaWUtZ2FsZXJpZS1hdC1raWFmLTIwMTg=",
-              partner: {
-                __typename: "Partner",
-                name: "DIE GALERIE",
-                href: "/die-galerie",
-                id: "UGFydG5lcjpkaWUtZ2FsZXJpZQ==",
-              },
-              name: "DIE GALERIE at KIAF 2018",
-              start_at: "2018",
-              city: null,
-              href: "/show/die-galerie-die-galerie-at-kiaf-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjI=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzptYXN0ZXJ3b3Jrcy1maW5lLWFydC1tYXN0ZXJ3b3Jrcy1maW5lLWFydC1hdC10ZXhhcy1jb250ZW1wb3JhcnktMjAxOA==",
-              partner: {
-                __typename: "Partner",
-                name: "Masterworks Fine Art",
-                href: "/masterworks-fine-art-gallery",
-                id: "UGFydG5lcjptYXN0ZXJ3b3Jrcy1maW5lLWFydA==",
-              },
-              name: "Masterworks Fine Art at Texas Contemporary 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/masterworks-fine-art-masterworks-fine-art-at-texas-contemporary-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjM=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzp0YW55YS1iYXh0ZXItY29udGVtcG9yYXJ5LXRhbnlhLWJheHRlci1jb250ZW1wb3JhcnktYXQtZmluZS1hcnQtYXNpYS0yMDE4",
-              partner: {
-                __typename: "Partner",
-                name: "Tanya Baxter Contemporary",
-                href: "/tanya-baxter-contemporary",
-                id: "UGFydG5lcjp0YW55YS1iYXh0ZXItY29udGVtcG9yYXJ5",
-              },
-              name: "Tanya Baxter Contemporary at Fine Art Asia 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/tanya-baxter-contemporary-tanya-baxter-contemporary-at-fine-art-asia-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjQ=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpnbGFkd2VsbC1hbmQtcGF0dGVyc29uLWdsYWR3ZWxsLWFuZC1wYXR0ZXJzb24tYXQtZmluZS1hcnQtYXNpYS0yMDE4",
-              partner: {
-                __typename: "Partner",
-                name: "Gladwell & Patterson",
-                href: "/gladwell-and-patterson",
-                id: "UGFydG5lcjpnbGFkd2VsbC1hbmQtcGF0dGVyc29u",
-              },
-              name: "Gladwell & Patterson at Fine Art Asia 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/gladwell-and-patterson-gladwell-and-patterson-at-fine-art-asia-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjU=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzphY2EtZ2FsbGVyaWVzLWFjYS1nYWxsZXJpZXMtYXQtc2VhdHRsZS1hcnQtZmFpci0yMDE4",
-              partner: {
-                __typename: "Partner",
-                name: "ACA Galleries",
-                href: "/aca-galleries",
-                id: "UGFydG5lcjphY2EtZ2FsbGVyaWVz",
-              },
-              name: "ACA Galleries at Seattle Art Fair 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/aca-galleries-aca-galleries-at-seattle-art-fair-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjY=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpmbC1icmFzd2VsbC1maW5lLWFydC1mbC1icmFzd2VsbC1maW5lLWFydC1hdC1zZWF0dGxlLWFydC1mYWlyLTIwMTg=",
-              partner: {
-                __typename: "Partner",
-                name: "F.L. Braswell Fine Art",
-                href: "/fl-braswell",
-                id: "UGFydG5lcjpmbC1icmFzd2VsbC1maW5lLWFydA==",
-              },
-              name: "F.L. Braswell Fine Art at Seattle Art Fair 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/fl-braswell-fine-art-fl-braswell-fine-art-at-seattle-art-fair-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjc=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpvcGVyYS1nYWxsZXJ5LW9wZXJhLWdhbGxlcnktYXQtYXJ0LWFzcGVuLTIwMTg=",
-              partner: {
-                __typename: "Partner",
-                name: "Opera Gallery",
-                href: "/opera-gallery",
-                id: "UGFydG5lcjpvcGVyYS1nYWxsZXJ5",
-              },
-              name: "Opera Gallery at Art Aspen 2018",
-              start_at: "2018",
-              city: null,
-              href: null,
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjg=",
-          },
-          {
-            node: {
-              id:
-                "U2hvdzpwZXRlci1oYXJyaW5ndG9uLWdhbGxlcnktcGV0ZXItaGFycmluZ3Rvbi1nYWxsZXJ5LWF0LW1hc3RlcnBpZWNlLWxvbmRvbi0yMDE4",
-              partner: {
-                __typename: "Partner",
-                name: "Peter Harrington Gallery",
-                href: "/peter-harrington-gallery",
-                id: "UGFydG5lcjpwZXRlci1oYXJyaW5ndG9uLWdhbGxlcnk=",
-              },
-              name: "Peter Harrington Gallery at Masterpiece London 2018",
-              start_at: "2018",
-              city: null,
-              href:
-                "/show/peter-harrington-gallery-peter-harrington-gallery-at-masterpiece-london-2018",
-              __typename: "Show",
-            },
-            cursor: "YXJyYXljb25uZWN0aW9uOjk=",
-          },
-        ],
-      },
-      id: "QXJ0aXN0OnBhYmxvLXBpY2Fzc28=",
-    },
-  },
-}
